requirements.txt
keras
pandas
streamlit==1.34
pillow
openai